refactor: move CLI source files into `src` directory and update all related paths
