+++ packages/cli/src/commands/generate.ts
@@ -5,11 +5,11 @@
  * This is effectively a semantic alias for the 'fix' command but focused on content generation.
  */
 
-import { DoctypeMapManager } from '../content/map-manager';
+import { DoctypeMapManager } from '../../../content/map-manager';
 import { AstAnalyzer, extractAnchors, DoctypeAnchor, CodeSignature } from '@doctypedev/core';
-import { Logger } from './logger';
-import { GenerateResult, GenerateOptions } from './types';
-import { detectDrift } from './drift-detector';
+import { Logger } from '../utils/logger';
+import { GenerateResult, GenerateOptions } from '../types';
+import { detectDrift } from '../services/drift-detector';
 import { existsSync, readFileSync } from 'fs';
 import { resolve, dirname } from 'path';
 import {
@@ -17,8 +17,8 @@ import {
   getMapPath,
   ConfigNotFoundError,
   InvalidConfigError,
-} from './config-loader';
-import { executeFixes } from './fix-orchestrator';
+} from '../services/config-loader';
+import { executeFixes } from '../orchestrators/fix-orchestrator';
 
 /**
  * Execute the generate command
@@ -88,7 +88,7 @@ export async function generateCommand(options: GenerateOptions): Promise<Generat
   logger.newline();
 
   // Resolve the root directory for source code
-  const codeRoot = config 
+  const codeRoot = config
     ? resolve(config.baseDir || process.cwd(), config.projectRoot)
     : dirname(mapPath);
 
@@ -103,10 +103,10 @@ export async function generateCommand(options: GenerateOptions): Promise<Generat
   // This enables "doctype generate" to work after "doctype init"
   const driftedIds = new Set(detectedDrifts.map(d => d.entry.id));
   const projectBase = config ? (config.baseDir || process.cwd()) : dirname(mapPath);
-  
+
   // We need to re-analyze files to get signatures if we find placeholders
   // Optimization: Only analyze if we find a placeholder
-  
+
   for (const entry of entries) {
     if (driftedIds.has(entry.id)) {
       continue;
@@ -121,26 +121,26 @@ export async function generateCommand(options: GenerateOptions): Promise<Generat
       const docContent = readFileSync(docFilePath, 'utf-8');
       const extractionResult = extractAnchors(docFilePath, docContent);
       const anchor = extractionResult.anchors.find((a: DoctypeAnchor) => a.id === entry.id);
-      
+
       if (anchor && anchor.content.includes('TODO: Add documentation for this symbol')) {
         logger.debug(`Found placeholder for ${entry.codeRef.symbolName}, marking for generation`);
-        
+
         // We need the current signature to generate content
         const codeFilePath = resolve(projectBase, entry.codeRef.filePath);
         if (existsSync(codeFilePath)) {
-           const signatures = analyzer.analyzeFile(codeFilePath);
-           const currentSignature = signatures.find((s: CodeSignature) => s.symbolName === entry.codeRef.symbolName);
-           
-           if (currentSignature) {
-             detectedDrifts.push({
-               entry,
-               currentSignature,
-               currentHash: currentSignature.hash!,
-               oldHash: entry.codeSignatureHash,
-               // No old signature needed as context since it's a new generation
-             });
-             driftedIds.add(entry.id);
-           }
+          const signatures = analyzer.analyzeFile(codeFilePath);
+          const currentSignature = signatures.find((s: CodeSignature) => s.symbolName === entry.codeRef.symbolName);
+
+          if (currentSignature) {
+            detectedDrifts.push({
+              entry,
+              currentSignature,
+              currentHash: currentSignature.hash!,
+              oldHash: entry.codeSignatureHash,
+              // No old signature needed as context since it's a new generation
+            });
+            driftedIds.add(entry.id);
+          }
         }
       }
     } catch (e) {
