+++ packages/cli/src/prompts/document-prompt.ts
